requirements.txt
pyro-ppl
torch